test(guard): keep native pretool tests within quality limit
